@@ -35,8 +35,8 @@
 		"postinstall:electron": "node scripts/postinstall.js"
 	},
 	"engines": {
-		"node": ">= 13.0.0",
-		"npm": ">= 6.0.0"
+		"node": ">= 14.0.0",
+		"npm": ">= 7.0.0"
 	},
 	"private": true,
 	"main": "index.js",
@@ -127,24 +127,24 @@
 		"node_modules"
 	],
 	"dependencies": {
-		"@ajsf/core": "0.1.5",
-		"@ajsf/material": "0.1.4",
-		"@angular/animations": "10.1.6",
-		"@angular/cdk": "10.2.5",
-		"@angular/common": "10.1.6",
-		"@angular/core": "10.1.6",
+		"@ajsf/core": "0.3.0-beta.4",
+		"@ajsf/material": "0.3.0-beta.4",
+		"@angular/animations": "11.0.8",
+		"@angular/cdk": "11.0.3",
+		"@angular/common": "11.0.8",
+		"@angular/core": "11.0.8",
 		"@angular/flex-layout": "10.0.0-beta.32",
-		"@angular/forms": "10.1.6",
-		"@angular/material": "10.2.5",
-		"@angular/material-moment-adapter": "10.2.5",
-		"@angular/platform-browser": "10.1.6",
-		"@angular/platform-browser-dynamic": "10.1.6",
-		"@angular/router": "10.1.6",
-		"@clr/angular": "4.0.3",
-		"@clr/core": "4.0.3",
-		"@clr/icons": "4.0.3",
-		"@clr/ui": "4.0.3",
-		"@fortawesome/angular-fontawesome": "^0.7.0",
+		"@angular/forms": "11.0.8",
+		"@angular/material": "11.0.3",
+		"@angular/material-moment-adapter": "11.0.3",
+		"@angular/platform-browser": "11.0.8",
+		"@angular/platform-browser-dynamic": "11.0.8",
+		"@angular/router": "11.0.8",
+		"@clr/angular": "^4.0.9",
+		"@clr/core": "4.0.9",
+		"@clr/icons": "4.0.9",
+		"@clr/ui": "4.0.9",
+		"@fortawesome/angular-fontawesome": "^0.8.1",
 		"@fortawesome/fontawesome-svg-core": "^1.2.32",
 		"@fortawesome/free-brands-svg-icons": "^5.15.1",
 		"@fortawesome/free-regular-svg-icons": "^5.15.1",
@@ -166,81 +166,88 @@
 		"bip32-path": "^0.4.2",
 		"body-parser": "^1.19.0",
 		"bootstrap": "^4.5.3",
-		"chart.js": "^2.9.3",
+		"chart.js": "^2.9.4",
 		"cors": "^2.8.5",
 		"crypto-js": "^4.0.0",
 		"echarts": "^4.9.0",
-		"electron-context-menu": "^2.3.0",
+		"electron-context-menu": "^2.4.0",
 		"eosjs": "^21.0.1-rc2",
 		"express": "^4.17.1",
 		"fcbuffer": "^2.2.2",
 		"fuse.js": "^3.4.6",
 		"hammerjs": "^2.0.8",
 		"isomorphic-fetch": "^3.0.0",
-		"jquery": "^3.5.1",
-		"keytar": "^5.6.0",
-		"lodash-es": "^4.17.15",
+		"jquery": "^3.5.0",
+		"keytar": "^7.3.0",
+		"lodash-es": "^4.17.20",
 		"lodash.filter": "4.6.0",
 		"lodash.get": "^4.4.2",
 		"lodash.map": "4.6.0",
 		"lodash.set": "4.3.2",
-		"lottie-web": "^5.7.3",
+		"lottie-web": "^5.7.5",
 		"menu": "^0.2.5",
 		"moment": "^2.29.1",
 		"ng-snotify": "^9.0.1",
-		"ngx-echarts": "^5.1.2",
+		"ngx-echarts": "^5.2.2",
 		"ngx-electron": "^2.2.0",
-		"ngx-json-viewer": "^2.4.0",
+		"ngx-json-viewer": "^3.0.2",
 		"ngx-lottie": "^6.4.0",
-		"ngx-markdown": "^10.1.1",
+		"ngx-markdown": "^11.0.1",
 		"ngx-order-pipe": "^2.1.1",
 		"ngx-pagination": "^5.0.0",
+		"node-abi": "^2.19.3",
 		"node-fetch": "^2.6.1",
-		"node-pre-gyp": "^0.15.0",
-		"node-sass": "^4.14.1",
+		"node-gyp": "^7.1.2",
+		"node-hid": "^2.1.1",
+		"node-pre-gyp": "^0.17.0",
 		"node-schedule": "^1.3.2",
-		"open": "^7.3.0",
+		"open": "^7.3.1",
 		"popper.js": "^1.16.1",
 		"portfinder": "^1.0.28",
-		"primeicons": "^4.0.0-rc.2",
-		"primeng": "^9.1.3",
+		"primeicons": "^4.1.0",
+		"primeng": "^11.1.0",
 		"quill": "^1.3.7",
-		"socket.io": "^2.3.0",
-		"socket.io-client": "^2.3.1",
+		"resize-observer-polyfill": "^1.5.1",
+		"sass": "^1.32.2",
+		"socket.io": "^3.0.5",
+		"socket.io-client": "^3.0.5",
 		"sudo-prompt": "^9.2.1",
-		"tar": "^6.0.5",
+		"tar": "^6.1.0",
 		"text-mask-addons": "3.8.0",
-		"tslib": "^2.0.3"
+		"text-mask-core": "5.1.2",
+		"tslib": "^1.14.1"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "0.1001.7",
-		"@angular/cli": "10.1.7",
-		"@angular/compiler": "10.1.6",
-		"@angular/compiler-cli": "10.1.6",
-		"@angular/language-service": "10.1.6",
+		"@angular-devkit/build-angular": "^0.1100.6",
+		"@angular/cli": "11.0.6",
+		"@angular/compiler": "11.0.8",
+		"@angular/compiler-cli": "11.0.8",
+		"@angular/language-service": "11.0.8",
 		"@electron/typescript-definitions": "8.8.0",
 		"@types/crypto-js": "4.0.1",
-		"@types/echarts": "4.8.1",
-		"@types/express": "4.17.8",
-		"@types/node": "14.11.8",
-		"@webcomponents/custom-elements": "^1.4.2",
+		"@types/echarts": "4.9.3",
+		"@types/express": "4.17.10",
+		"@types/node": "14.14.20",
+		"@webcomponents/custom-elements": "^1.4.3",
 		"codelyzer": "^6.0.1",
-		"core-js": "^3.6.5",
-		"electron": "9.3.2",
-		"electron-builder": "22.9.1",
+		"core-js": "^3.8.2",
+		"electron": "^11.2.0",
+		"electron-builder": "22.10.4",
+		"electron-rebuild": "^2.3.4",
 		"electron-builder-notarize": "^1.2.0",
 		"electron-reload": "^1.5.0",
-		"eslint": "^7.11.0",
+		"eslint": "^7.17.0",
+		"html-webpack-plugin": "^4.5.1",
 		"npm-run-all": "^4.1.5",
 		"protractor": "^7.0.0",
 		"rxjs": "^6.6.3",
 		"rxjs-compat": "^6.6.3",
 		"text-encoding-shim": "^1.0.5",
-		"ts-node": "^9.0.0",
+		"ts-node": "^9.1.1",
 		"typescript": "4.0.3",
-		"wait-on": "^5.2.0",
-		"webpack": "^4.44.2",
-		"webpack-dev-server": "^3.11.0",
-		"zone.js": "^0.11.1"
+		"wait-on": "^5.2.1",
+		"webpack": "^4.46.0",
+		"webpack-dev-server": "^3.11.1",
+		"zone.js": "^0.10.3"
 	}
 }
