@@ -1,7 +1,7 @@
 const fs = require('fs');
 
 // patch webpack browser config for electron
-const f_angular = 'node_modules/@angular-devkit/build-angular/src/angular-cli-files/models/webpack-configs/browser.js';
+const f_angular = 'node_modules/@angular-devkit/build-angular/src/webpack/configs/browser.js';
 const data = fs.readFileSync(f_angular, 'utf8');
 if (!data) {
     return console.log('failed to read:' + f_angular);
